#25 Feat: Implement Plugin Recognition in Core ( closes #25 )
