+++ SOK_Graph_Project/flask/app.py
@@ -11,17 +11,22 @@ sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
 project_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
 
-from json_data_source.json_db.json_db import JSONRepository
-from core.service.use_cases.plugin_recognition import recognize_plugin
 from core.service.use_cases.tree_view import TreeView
 from core.service.use_cases.bird_view import BirdView
 from flask import Flask, render_template, request
-from csv_data_source.csv_db.csv_db import CsvDb
+from plugins.csv_data_source.csv_db.csv_db import CsvDb
+#from plugins.plugin_loader import load_plugins
+from core.service.use_cases.plugin_recognition import PluginService
 
+plugin_service = PluginService()
+
+def get_plugins():
+    return plugin_service.load_plugins()
 
 app = Flask(__name__)
 app.config['APP_NAME'] = "Graph Visualizer"
 app.logger.setLevel(logging.INFO)
+plugins = get_plugins()
 
 # Allow Flask to load templates both from flask/templates and core/templates
 flask_templates = os.path.join(project_root, 'flask', 'templates')
@@ -37,10 +42,11 @@ app.jinja_loader = ChoiceLoader([
 def index():
     mode = request.args.get("mode", "separate_files")
     dataset = request.args.get("dataset", "acyclic")
-    user_choice = request.args.get("type", "simple")  # "simple" or "block"
+    selected_data_source = request.args.get("data_source", "csv")
+    selected_visualizer = request.args.get("type", "simple_visualizer")
 
     base_dir = os.path.abspath(
-        os.path.join(project_root, 'csv_data_source', 'csv_data')
+        os.path.join(project_root, 'plugins', 'csv_data_source', 'csv_data')
     )
 
     if mode == "single_file":
@@ -62,23 +68,38 @@ def index():
             edges_path=edges_path
         )
 
-    # Load graph from CSV
-    graph = csv_db.load()
-
     # Load graph from JSON
-    repository = JSONRepository("test_graph.json")
-    json_graph = repository.read_from_file()
-    json_graph_dict = repository.graph_to_dict(json_graph)
+    # repository = JSONRepository("test_graph.json")
+    # json_graph = repository.read_from_file()
+    # json_graph_dict = repository.graph_to_dict(json_graph)
 
     # if json_graph_dict is None:
     #     app.logger.info("Failed to load graph from file.")
     # else:
     #     app.logger.info("Graph structure:\n%s", json.dumps(json_graph_dict, indent=2, ensure_ascii=False))
 
-    # Main visualization
+    if selected_data_source:
+        ds_plugin = plugins["data_source"].get(selected_data_source)
+        if ds_plugin:
+            graph = ds_plugin.load()
+        else:
+            graph = csv_db.load()
+    else:
+        graph = csv_db.load()
+
+    graph_dict = csv_db.to_dict(graph)
+
     app.config["GRAPH"] = graph
-    user_choice = request.args.get("type")  # simple ili block
-    visualizer = recognize_plugin(user_choice)
+
+    visualizer = plugins["visualization"].get(selected_visualizer)
+
+    if visualizer is None:
+        if plugins["visualization"]:
+            visualizer = next(iter(plugins["visualization"].values()))
+            print(f"[app] Plugin '{selected_visualizer}' nije pronađen, koristim prvi dostupan: '{visualizer.identifier()}'")
+        else:
+            return "No visualization plugins loaded"
+
     graph_html = visualizer.visualize(graph)
 
     # Bird view
@@ -95,11 +116,13 @@ def index():
         graph_html=graph_html,
         bird_view=bird_view_html,
         tree_view_html=tree_view_html,
-        selected_visualizer=user_choice,
         selected_mode=mode,
-        selected_dataset=dataset
+        selected_dataset=dataset,
+        plugins=plugins["visualization"],
+        data_plugins=plugins["data_source"],
+        selected_visualizer=selected_visualizer,
+        selected_data_source=selected_data_source,
     )
-
 
 if __name__ == "__main__":
     app.run(debug=True)
