+++ SOK_Graph_Project/plugins/simple_visualizer/simple_visualizer.py
